+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.2.0
+### Added
+The String class and two methods:
+
+- startsWith
+- endsWith
+
+A README, finally!
+
 ## 0.1.3
 ### Added
 The array class and two methods: